+++ packages/atomic/src/components.d.ts
@@ -302,6 +302,10 @@ export namespace Components {
         "maxNumberOfParts": number;
     }
     interface AtomicResultQuickview {
+        /**
+          * The maximum preview size to retrieve, in bytes. By default, the full preview is retrieved.
+         */
+        "maximumPreviewSize"?: number | undefined;
     }
     interface AtomicResultTemplate {
         /**
@@ -968,6 +972,10 @@ declare namespace LocalJSX {
         "maxNumberOfParts"?: number;
     }
     interface AtomicResultQuickview {
+        /**
+          * The maximum preview size to retrieve, in bytes. By default, the full preview is retrieved.
+         */
+        "maximumPreviewSize"?: number | undefined;
     }
     interface AtomicResultTemplate {
         /**